requirements.txt
fastapi
uvicorn
jinja2
watchdog>=3.0.0
python-frontmatter>=1.0.0
requests
pydantic>=2.9.0,<2.11.0
pydantic-settings>=2.5.0,<2.6.0
passlib[bcrypt]
# Avoid bcrypt 4.x incompat with older passlib handlers
bcrypt>=3.2,<4.0
python-jose
python-multipart
python-dotenv
# System monitoring for health checks
psutil>=5.9.0
sentence-transformers>=2.2.2
# Re-ranking and sparse search
rank-bm25>=0.2.2
scikit-learn>=1.3.0
discord.py>=2.3.2
aiohttp>=3.9.0
# Image and document processing
Pillow>=10.0.0
pytesseract>=0.3.10
PyPDF2>=3.0.0
python-magic>=0.4.27
psutil

# Optional lightweight ASR backend (set transcriber=vosk)
#vosk>=0.3.45

# Web content extraction
playwright>=1.40.0
beautifulsoup4>=4.12.0
html2text>=2020.1.16
readability-lxml>=0.8.1
